Add AbstractSlashCommandResponseParms.
