+++ slack-base/src/main/java/com/hubspot/slack/client/methods/params/chat/AbstractChatPostMessageParams.java
@@ -1,46 +1,13 @@
 package com.hubspot.slack.client.methods.params.chat;
 
-import static com.hubspot.slack.client.models.TopLevelMessageResponseType.IN_CHANNEL;
-
-import java.util.Optional;
-
-import org.immutables.value.Value.Check;
-import org.immutables.value.Value.Default;
 import org.immutables.value.Value.Immutable;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.PropertyNamingStrategy.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import com.google.common.base.Preconditions;
-import com.google.common.base.Strings;
 import com.hubspot.immutables.style.HubSpotStyle;
-import com.hubspot.slack.client.models.TopLevelMessageResponseType;
 
 @Immutable
 @HubSpotStyle
 @JsonNaming(SnakeCaseStrategy.class)
-public abstract class AbstractChatPostMessageParams implements MessageParams {
-  @JsonProperty("channel")
-  public abstract String getChannelId();
-  @Default
-  public TopLevelMessageResponseType getResponseType() {
-    return IN_CHANNEL;
-  }
-  public abstract Optional<String> getText();
-  public abstract Optional<String> getThreadTs();
-  public abstract Optional<String> getUsername();
-  public abstract Optional<Boolean> getAsUser();
-  public abstract Optional<String> getIconEmoji();
-  public abstract Optional<String> getIconUrl();
-  public abstract Optional<Boolean> getLinkNames();
-  public abstract Optional<Boolean> getUnfurlLinks();
-  public abstract Optional<Boolean> getUnfurlMedia();
-  public abstract Optional<Boolean> getReplyBroadcast();
-
-  @Check
-  public void check() {
-    Preconditions.checkState((getText().isPresent() && !Strings.isNullOrEmpty(getText().get())) ||
-            !getAttachments().isEmpty(),
-        "Must include text if not providing attachments");
-  }
+public abstract class AbstractChatPostMessageParams extends AbstractChatMessageParams {
 }
