+++ slack-base/src/main/java/com/hubspot/slack/client/methods/params/chat/AbstractSlashCommandResponseParams.java
@@ -1,0 +1,14 @@
+package com.hubspot.slack.client.methods.params.chat;
+
+import static com.hubspot.slack.client.models.TopLevelMessageResponseType.EPHEMERAL;
+
+import org.immutables.value.Value.Default;
+
+import com.hubspot.slack.client.models.TopLevelMessageResponseType;
+
+public abstract class AbstractSlashCommandResponseParams extends AbstractChatMessageParams {
+  @Default
+  public TopLevelMessageResponseType getResponseType() {
+    return EPHEMERAL;
+  }
+}
